requirements.txt
fastapi
uvicorn[standard]
playwright
pydantic
python-dotenv